requirements.txt
Flask>=1.0
psycopg2-binary>=2.7
gunicorn>=19.9
flask-login>=0.4.0
Flask-WTF>=0.14
WTForms-Components>=0.10.4
passlib>=1.7.2
colour>=0.1.5